Move VaporJSONServiceBridge to generated library
