+++ example/test.sh
@@ -8,9 +8,12 @@ fi
 swift build -c release
 .build/release/Server &
 SERVER=$!
+function finally {
+  kill $SERVER
+}
+trap finally EXIT
 sleep 0.1
 .build/release/Client
 cd TSClient
 npm run run
-kill $SERVER
